Read configuration from multiple SSM parameters
Use a separate stack for configuring endpoints the
status-checker should check.

The new stack creates an SSM parameter containing the
configuration and setup the Event rule that schedules
when the status-checker should be invoked.

The lambda expect now to be invoked with an event that
contains the parameter to get from SSM.

Fixes: https://github.com/frodeaa/status-checker/issues/21